Use uint64_t instead of size_t

+++ src/tools/spin/spinquic.cpp
@@ -557,7 +557,7 @@ void Spin(LockableVector<HQUIC>& Connections, std::vector<HQUIC>* Listeners = nu
                 if (Stream == nullptr) continue;
                 auto BytesRemaining = MsQuic->GetContext(Stream);
                 if (BytesRemaining != nullptr && GetRandom(10) == 0) {
-                    auto BytesConsumed = GetRandom((size_t)BytesRemaining);
+                    auto BytesConsumed = GetRandom((uint64_t)BytesRemaining);
                     MsQuic->StreamReceiveComplete(Stream, BytesConsumed);
                     MsQuic->SetContext(Stream, (void*)((size_t)BytesRemaining - BytesConsumed));
                 } else {
